+++ docs/.vitepress/theme/index.js
@@ -1,6 +1,7 @@
 import DefaultTheme from 'vitepress/theme'
 import VersionTag from './components/VersionTag.vue'
 import RomanNumerals from './components/RomanNumerals.vue'
+import ChinaNumerals from './components/ChinaNumerals.vue'
 import '../../../lib/index.umd'
 import './index.scss'
 
@@ -10,5 +11,6 @@ export default {
     // 注册全局组件
     app.component('VersionTag', VersionTag)
     app.component('RomanNumerals', RomanNumerals)
+    app.component('ChinaNumerals', ChinaNumerals)
   }
 }